Refactored repos by separating Comment into its own Repository
