+++ SarasTreasures/Models/Data/IStoryRepository.cs
@@ -5,10 +5,6 @@ namespace SarasTreasures.Models
 {
     public interface IStoryRepository
     {
-        // -----------
-        //  S T O R Y
-        // -----------
-
         // Create
         void AddStory(Story story);
         // Retrieve 
@@ -19,19 +15,5 @@ namespace SarasTreasures.Models
         void UpdateStory(Story story);
         // Delete
         void DeleteStory(int id);
-
-        // ---------------
-        //  C O M M E N T
-        // ---------------
-
-        // Create
-        void AddComment(Comment comment);
-        // Retrieve 
-        IQueryable<Comment> Comments { get; }
-        Comment GetCommentByID(int id);
-        // Update
-        void UpdateComment(Comment comment);
-        // Delete
-        void DeleteComment(int id);
     }
 }
